rename: arquivo precos.entity @JoinColumn
